refactor: moved PySRSequenceRegressor to ssr.py
