+++ pysr/__init__.py
@@ -7,7 +7,8 @@ from . import sklearn_monkeypatch
 from .deprecated import best, best_callable, best_row, best_tex, install, pysr
 from .export_jax import sympy2jax
 from .export_torch import sympy2torch
-from .sr import PySRRegressor, PySRSequenceRegressor
+from .sr import PySRRegressor
+from .ssr import PySRSequenceRegressor
 
 # This file is created by setuptools_scm during the build process:
 from .version import __version__
@@ -20,6 +21,7 @@ __all__ = [
     "sympy2torch",
     "install",
     "PySRRegressor",
+    "PySRSequenceRegressor",
     "best",
     "best_callable",
     "best_row",
